separated view and controller components
