+++ src/main.jsx
@@ -4,13 +4,13 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import App from "./App";
 import "./index.css";
 import ServerClientAnimation from "./pages/ServerClient/ServerClient.view";
-import InsertionSortView from "./pages/InsertionSort/InsertionSort.view";
+import InsertionSortController from "./pages/InsertionSort/InsertionSort.controller";
 
 const router = createBrowserRouter([
   { path: "/", element: <App /> },
   { path: "/bubble", element: null },
   { path: "/selection", element: null },
-  { path: "/insertion", element: <InsertionSortView /> },
+  { path: "/insertion", element: <InsertionSortController /> },
   { path: "/quick", element: null },
   { path: "/merge", element: null },
   { path: "/linear", element: null },
